refactor: move robot_ai to application layer
- Copy robot_ai module from simulation to application layer
- Remove old robot_ai directory from simulation
- Update all imports across the codebase using sed replacement
- Robot AI logic now properly in application layer
- Major clean architecture reorganization step complete
