+++ src/application/robot_ai/mod.rs
@@ -1,1 +1,7 @@
-// TODO: Move robot_ai here
+pub mod behavior;
+pub mod behaviors;
+pub mod executor;
+pub mod pathfinding;
+pub mod robot;
+pub mod types;
+pub mod utils;
