+++ src/application/robot_ai/types.rs
