+++ src/application/robot_ai/utils.rs
