+++ src/simulation/mod.rs
@@ -1,6 +1,5 @@
 pub mod entities;
 pub mod map;
-pub mod robot_ai;
 
 #[cfg(test)]
 mod tests;
