--- src/simulation/robot_ai/mod.rs
@@ -1,7 +1,0 @@
-pub mod behavior;
-pub mod behaviors;
-pub mod executor;
-pub mod pathfinding;
-pub mod robot;
-pub mod types;
-pub mod utils;
